Mejora del ejercicio de java del examen
