--- UA1/tareas/dev_0/Tarea 3 - Programación de procesos en Java (II).java
@@ -1,22 +1,0 @@
-public class ProcesoPrincipal {
-    public static void main(String[] args) {
-        if (args.length < 1) {
-            System.exit(1); 
-        }
-        
-        try {
-
-            int num = Integer.parseInt(args[0]);
-            
-
-            if (num < 0) {
-                System.exit(3);
-            } else {
-                System.exit(0);
-            }
-        } catch (NumberFormatException e) {
-            
-            System.exit(2);
-        }
-    }
-}
